requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4            # via virtualenv
boto3==1.14.60            # via django-s3-storage
botocore==1.17.60         # via boto3, s3transfer
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via pip-tools
distlib==0.3.1            # via virtualenv
django-debug-toolbar==2.2  # via -r requirements.in
django-jet2==1.0.12       # via -r requirements.in
django-s3-storage==0.13.3  # via -r requirements.in
django-storages==1.10     # via -r requirements.in
django==2.2.16            # via -r requirements.in, django-debug-toolbar, django-jet2, django-s3-storage, django-storages, djangorestframework
djangorestframework==3.11.1  # via -r requirements.in
docutils==0.15.2          # via botocore
filelock==3.0.12          # via virtualenv
identify==1.5.1           # via pre-commit
importlib-metadata==1.7.0  # via pre-commit, virtualenv
importlib-resources==3.0.0  # via pre-commit, virtualenv
jmespath==0.10.0          # via boto3, botocore
mysql-connector-python==8.0.21  # via -r requirements.in
nodeenv==1.5.0            # via pre-commit
pip-tools==5.3.1          # via -r requirements.in
pre-commit==2.7.1         # via -r requirements.in
protobuf==3.13.0          # via mysql-connector-python
python-dateutil==2.8.1    # via botocore
python-dotenv==0.14.0     # via -r requirements.in
pytz==2020.1              # via django
pyyaml==5.3.1             # via pre-commit
s3transfer==0.3.3         # via boto3
six==1.15.0               # via pip-tools, protobuf, python-dateutil, virtualenv
sqlparse==0.3.1           # via django, django-debug-toolbar
toml==0.10.1              # via pre-commit
urllib3==1.25.10          # via botocore
virtualenv==20.0.31       # via pre-commit
zipp==3.1.0               # via importlib-metadata, importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
